kselfservice/backend/requirements.txt
fastapi
uvicorn
pydantic
PyYAML
requests
python-dotenv
casbin
